PROYECTO-TRIPLETEN/requirements.txt
plotly-express
pandas
streamlit

